+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new method on JWT `unknownKid()` to generate a random value for the `kid`
   header. This normally causes the `kid` to not match any previously generated
   key
+- new method on JWT `withIssuer` to quickly update the `iss` claim on the current
+  JWT
 
 ## [0.9.3] - 2024-06-13